add test for build message
