+++ selfdrive/proclogd/tests/.gitignore
@@ -1,1 +1,1 @@
-test_parser
+test_proclog
